+++ src/TrashLib/Radarr/CustomFormat/Guide/IRadarrGuideService.cs
@@ -1,6 +1,8 @@
+using TrashLib.Radarr.CustomFormat.Models;
+
 namespace TrashLib.Radarr.CustomFormat.Guide;
 
 public interface IRadarrGuideService
 {
-    IEnumerable<string> GetCustomFormatJson();
+    IEnumerable<CustomFormatData> GetCustomFormatData();
 }