+++ src/TrashLib/Radarr/CustomFormat/ICustomFormatLister.cs
@@ -1,0 +1,6 @@
+namespace TrashLib.Radarr.CustomFormat;
+
+public interface ICustomFormatLister
+{
+    void ListCustomFormats();
+}